Verify memory reading of ingamenames before 1st move selection
In very rare cases, the memory address may be incorrect.  When that occurs, this change ensures the game crashes before the first move selection, which typically results in a replayed match.
Previously the game would continue to play, with switches not being recognized by the engine, causing big gameplay and overlay issues.